requirements.base.txt
pandas==2.2.3
scikit-learn==1.6.1 
joblib==1.4.2 
mlflow==2.21.1
python-dotenv